Added 2d array grid to state of Grid component
The 2d array grid in state holds the contents of the displayed grid. This required refactoring of the generateGameBoard to directly use the object when rendering the game board. The class GridCell holds the information for each gridcell and includes a list of objects that are currently on the cell. These are meant to be either the player or an audio sprite.